Move race timer logic into separate component
