+++ frontend/src/components/pages/runner-details/RunnerDetailsRaceDetails.tsx
@@ -1,50 +1,14 @@
-import React, {useEffect, useMemo, useState} from "react";
+import React, {useMemo} from "react";
 import {Race} from "../../../types/Race";
 import Util from "../../../util/Util";
-import {app} from "../../App";
+import RaceTimer from "../../misc/RaceTimer";
 
 const RunnerDetailsRaceDetails: React.FunctionComponent<{
     race: Race
 }> = ({race}) => {
-    // The current race time
-    const [raceTime, setRaceTime] = useState(0);
-
     // The race total duration formatted to be human-readable
     const formattedRaceDuration = useMemo(() => {
         return Util.formatMsAsDuration(race.duration * 1000);
-    }, [race]);
-
-    const formattedRaceTime = useMemo(() => {
-        return Util.formatMsAsDuration(raceTime);
-    }, [raceTime]);
-
-    useEffect(() => {
-        const updateRaceTime = () => {
-            const raceStartMs = new Date(race.startTime).getTime();
-            const nowMs = (new Date()).getTime() + (app.state.serverTimeOffset * 1000);
-
-            const raceTime = nowMs - raceStartMs;
-
-            if (raceTime < 0) {
-                setRaceTime(0);
-                return;
-            }
-
-            if (raceTime > race.duration * 1000) {
-                setRaceTime(race.duration * 1000);
-                return;
-            }
-
-            setRaceTime(raceTime);
-        };
-
-        const interval = window.setInterval(updateRaceTime, 1000);
-
-        updateRaceTime();
-
-        return () => {
-            window.clearInterval(interval);
-        }
     }, [race]);
 
     return (
@@ -54,7 +18,7 @@ const RunnerDetailsRaceDetails: React.FunctionComponent<{
 
                 <p>{race.name}</p>
 
-                <p><b>{formattedRaceTime}</b> / {formattedRaceDuration}</p>
+                <p><b><RaceTimer race={race}/></b> / {formattedRaceDuration}</p>
 
                 <p>
                     Distance tour : <strong>{race.lapDistance} m</strong> {race.initialDistance > 0 &&
